CI: Reduce log output
This commit makes it such that circle CI only shows the module whose
tests it is currently running in the log, unless a test fails. For each
failing test, it will display all of the log output. This is done to make
our log output far more scrollable. We lose no information in debugging.